Refactorización.
Se reorganizan las clases del paquete validations agregando un nuevo
paquete labels para las etiquetas.
